Using the Deconstruction

+++ LearningCSharp7/Source/Learn.TuplesDemo/Subject.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
 
-        public decimal Marks { get; set; }
+        public float Marks { get; set; }
     }
 
 }
